Update radio runtime for improved efficiency
The radio runtime was updated to streamline its operations and increase efficiency. Changes include the modification of transmitted signals in the `test_runtime.rs` from all max signals `[255, 255, 255, 255]` to midway signals `[127, 127, 127, 127]` in order to test more common signal values, and the reduction of sleep time from 10 seconds to 1 for faster execution.

Moreover, in the `runtime.rs`, unnecessary parts of code were removed and replaced with a more straightforward and efficient implementation. This comprises the removal of unused `Sender`, `Bytes`, and some channels, with the replacement of a single, simpler `Receiver<u8>`. Additionally, the codebase in the `rx()` function was cleaned up for more straightforward reading and receiving data.

These changes improve the overall efficiency of the program and make it easier to understand and maintain in the future.